Backend/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
wheel==0.45.1
azure-cosmos==4.9.0
websockets==15.0.1
pydantic-settings==2.9.1
openai==1.76.2
httpx==0.28.1
azure-eventhub==5.15.0
azure-identity==1.21.0
aiohttp==3.11.18